Add files via upload
The SMS Spam Collection Dataset is a public dataset containing SMS messages that have 
been labelled as either spam or legitimate. The UCI Machine Learning Repository provides a 
dataset comprising 5,572 SMS messages in English. Out of these messages, 4,827 are categorized 
as legitimate and 747 as spam. The dataset was initially compiled by Tiago A. Almeida and José 
María Gómez Hidalgo and is available for download. In this dataset, each of the row represents a 
single SMS message and each column corresponds to a word in the vocabulary.
The classification problem for the selected "Spam SMS" dataset is to classify a given SMS 
message as either spam or not spam (ham). The dataset is suitable for this problem because it 
consists of a large number of text messages, both spam and ham, that have been labelled by humans. 
This labelled dataset can be used to train and test a supervised machine learning model to classify 
new, unseen messages as either spam or ham